+++ avahi.h
@@ -24,6 +24,6 @@ struct service_tab {
 int avahi_browse(struct service_tab *table,int tabsize,char const *service_name);
 void avahi_free_service_table(struct service_tab *table,int tabsize);
 
-int avahi_start(char const *service_name,char const *service_type,int service_port,char const *dns_name,int base_address,char const *description,void *,int *);
+int avahi_start(char const *service_name,char const *service_type,int service_port,char const *dns_name,int base_address,char const *description,void *,size_t *);
 #define AVAHI_H 1
 #endif
